+++ tool/src/main/java/org/zaproxy/addon/migt/At_Hash_update.java
@@ -1,22 +1,3 @@
-/*
- * Zed Attack Proxy (ZAP) and its related class files.
- *
- * ZAP is an HTTP/HTTPS proxy for assessing web application security.
- *
- * Copyright 2024 The ZAP Development Team
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package org.zaproxy.addon.migt;
 
 import java.security.MessageDigest;
@@ -133,6 +114,7 @@ public class At_Hash_update extends Module {
             j.private_key_pem = sign_key;
             new_id_token = j.build();
         } catch (ParsingException e) {
+            System.out.println("Error position is At_Hash_update 1");
             applicable = false;
             return;
         }