+++ tool/src/main/java/org/zaproxy/addon/migt/Operation.java
@@ -1,22 +1,3 @@
-/*
- * Zed Attack Proxy (ZAP) and its related class files.
- *
- * ZAP is an HTTP/HTTPS proxy for assessing web application security.
- *
- * Copyright 2024 The ZAP Development Team
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package org.zaproxy.addon.migt;
 
 import static org.zaproxy.addon.migt.Tools.buildStringWithVars;
@@ -465,6 +446,7 @@ public class Operation extends Module {
                 if (!applicable) return;
             } catch (ParsingException e) {
                 applicable = false;
+                System.out.println("Error position is Operation 1");
                 e.printStackTrace();
                 return;
             }
@@ -480,6 +462,7 @@ public class Operation extends Module {
                 } catch (ParsingException e) {
                     e.printStackTrace();
                     applicable = false;
+                    System.out.println("Error position is Operation 1");
                     return;
                 }
             }
@@ -492,6 +475,7 @@ public class Operation extends Module {
                 } catch (ParsingException e) {
                     e.printStackTrace();
                     applicable = false;
+                    System.out.println("Error position is Operation 1");
                     return;
                 }
             }
@@ -537,6 +521,7 @@ public class Operation extends Module {
 
         } catch (ParsingException | PatternSyntaxException e) {
             applicable = false;
+            System.out.println("Error position is Operation 1");
             e.printStackTrace();
             return;
         }