+++ tool/src/main/java/org/zaproxy/addon/migt/Tools.java
@@ -1,22 +1,3 @@
-/*
- * Zed Attack Proxy (ZAP) and its related class files.
- *
- * ZAP is an HTTP/HTTPS proxy for assessing web application security.
- *
- * Copyright 2024 The ZAP Development Team
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package org.zaproxy.addon.migt;
 
 import com.google.gson.JsonElement;
@@ -37,6 +18,7 @@ import org.json.JSONObject;
 
 /** Class with methods to process messages and execute tests */
 public class Tools {
+
     /**
      * This function execute a list of checks over a message, returning true if all the checks are
      * successful
@@ -49,6 +31,7 @@ public class Tools {
     public static boolean executeChecks(
             List<Check> checks, HTTPReqRes message, boolean isRequest, List<Var> vars)
             throws ParsingException {
+        System.out.println("eseguito executeChecks");
         for (Check c : checks) {
             if (!c.execute(message, isRequest, vars)) {
                 return false;