+++ tool/src/main/java/org/zaproxy/addon/migt/samlraider/gui/SamlMain.java
@@ -7,8 +7,6 @@ import org.zaproxy.addon.migt.samlraider.application.SamlTabController;
 public class SamlMain extends JPanel {
 
     private static final long serialVersionUID = 1L;
-    //  private ITextEditor textEditorAction;
-    //  private ITextEditor textEditorInformation;
     private transient SamlTabController controller;
     private SamlPanelAction panelAction;
     private SamlPanelInfo panelInformation;