+++ tool/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## [0.0.1] - 2024-10-01
 ### Added
-- Addon creato nella cartela zap-extension
+- Addon created in separated project